+++ e2e/tests/utils/mock.ts
@@ -1,40 +1,31 @@
 import { Page } from "@playwright/test";
-// import { ProcessingItemType } from "../../../src/types";
-// let mockList: ProcessingItemType[] = [];
 
 export async function mockAPI(page: Page) {
   await page.route("*/**/check", async (route) => {
     const json = {
       noToken: false,
       output: "",
+      parameters: {
+        api: {
+          ENABLE_BEETS: "true",
+          ENABLE_PLEX_UPDATE: "true",
+          PLEX_URL: "http://plex.url",
+          PLEX_LIBRARY: "3",
+          PLEX_TOKEN: "abc-plex-token-xyz",
+          PLEX_PATH: "/fodler/to/plex/music",
+          ENABLE_GOTIFY: "true",
+          GOTIFY_URL: "http://gotify.url",
+          GOTIFY_TOKEN: "abc-gotify-token-xyz",
+          TIDARR_VERSION: "0.0.0",
+          TIDARR_REPO_URL: "cstaelen/tidarr",
+        },
+        app: {
+          REACT_APP_TIDAL_SEARCH_TOKEN: "CzET4vdadNUFQ5JU",
+          REACT_APP_TIDAL_COUNTRY_CODE: "CA",
+          REACT_APP_TIDARR_SEARCH_URL: "http://www.tidal.com/album/",
+        },
+      },
     };
     await route.fulfill({ json });
   });
-
-  // await page.route("*/**/list", async (route) => {
-  //   const json = mockList;
-  //   await route.fulfill({ json });
-  // });
-
-  // await page.route("*/**/remove", async (route) => {
-  //   mockList = [];
-  //   const json = mockList;
-  //   await route.fulfill({ json });
-  // });
-
-  // await page.route("*/**/save", async (route) => {
-  //   const json: ProcessingItemType = {
-  //     id: "77610756",
-  //     artist: "Nirvana",
-  //     title: "Nevermind",
-  //     type: "album",
-  //     status: "processing",
-  //     loading: true,
-  //     error: false,
-  //     url: "http://www.tidal.com/album/77610756",
-  //     output: "\r\n=== Tidal-DL ===",
-  //   };
-  //   mockList.push(json);
-  //   await route.fulfill({ json });
-  // });
 }